core/router: moved router to core
